Step will now churn process nops. Started working on memory map
I need to rethink how I'm managing supporting data structures. C++ is such a pain when it comes to declarations. I wwanted to forward declare a bunch of stuff in the header and define them and supporting types in the implementation file, but I'd need to have a mish mash of anonymouse and named namespaces to declare things in the right order.

That's my bad, I'll rethink that.